+++ src/module.vue
@@ -7,14 +7,31 @@ const api = useApi();
 
 const rapidoc = ref(null);
 const rapidocus_schema = ref("");
+var timer;
 
-const show_header = ref(false);
+// General
+const display_mode = ref("focused");
+const display_modes = ref(["view", "read", "focused"]);
+const allow_try = ref(true);
+
+// Header
+const show_header = ref(true);
+const allow_spec_url_load = ref(false);
+const allow_spec_file_load = ref(false);
+
+// Navigation
+const allow_search = ref(true);
+const allow_advanced_search = ref(true);
 const show_info = ref(false);
 const allow_server_selection = ref(false);
-
+const allow_authentication = ref(false);
+const persist_auth = ref(false);
+const font_spacing = ref("default");
+const font_spacings = ref(["compact", "default", "relaxed"]);
+const font_size = ref("default");
+const font_sizes = ref(["default", "large", "largest"]);
 const nav_active_item_marker = ref("left-bar");
 const nav_active_items_marker = ref(["left-bar", "colored-block"]);
-
 const show_method_in_nav_bar = ref("as-colored-block");
 const show_methods_in_nav_bar = ref([
   "false",
@@ -22,13 +39,7 @@ const show_methods_in_nav_bar = ref([
   "as-colored-text",
   "as-colored-block",
 ]);
-
-const display_mode = ref("focused");
-const display_modes = ref(["view", "read", "focused"]);
-
-const persist_auth = ref(false);
-
-var timer;
+//
 
 onMounted(async () => {
   const { data } = await api.get("/server/specs/oas");
@@ -44,15 +55,15 @@ watch(rapidocus_schema, (schema) => {
 });
 
 const colorScheme = window.matchMedia("(prefers-color-scheme: dark)");
-const mode = ref(colorScheme.matches ? "dark" : "light");
-
 colorScheme.addEventListener("change", (event) => {
   mode.value = event.matches ? "dark" : "light";
 });
 
+const mode = ref(colorScheme.matches ? "dark" : "light");
 const colors = computed(() => {
   if (mode.value === "dark") {
     return {
+      header: "#0d1117",
       bg: "#161b22",
       text: "#ffffff",
       primary: "#8866ff",
@@ -61,6 +72,7 @@ const colors = computed(() => {
     };
   } else {
     return {
+      header: "#ffffff",
       bg: "#ffffff",
       text: "#21262e",
       primary: "#8866ff",
@@ -77,32 +89,60 @@ const colors = computed(() => {
       v-if="rapidocus_schema"
       ref="rapidoc"
       style="height: 100%; width: 100%"
-      :show-method-in-nav-bar="show_method_in_nav_bar"
-      :nav-active-item-marker="nav_active_item_marker"
-      :show-header="show_header"
-      :show-info="show_info"
-      :allow-server-selection="allow_server_selection"
-      allow-authentication="false"
-      allow-search="true"
-      allow-advanced-search="true"
-      allow-try="true"
       :render-style="display_mode"
       :theme="mode"
+      :show-header="show_header"
+      :show-info="show_info"
+      :allow-spec-file-load="allow_spec_file_load"
+      :allow-spec-url-load="allow_spec_url_load"
+      :allow-server-selection="allow_server_selection"
+      :allow-authentication="allow_authentication"
+      :allow-search="allow_search"
+      :allow-advanced-search="allow_advanced_search"
+      :show-method-in-nav-bar="show_method_in_nav_bar"
+      :nav-active-item-marker="nav_active_item_marker"
+      :nav-item-spacing="font_spacing"
+      :font-size="font_size"
+      :allow-try="allow_try"
+      :header-color="colors.header"
       :bg-color="colors.bg"
       :primary-color="colors.primary"
       :nav-bg-color="colors.navBg"
       :nav-text-color="colors.navText"
       :persist-auth="persist_auth"
-    ></rapi-doc>
+      allow-schema-description-expand-toggle="true"
+    >
+      <div slot="header"></div>
+      <div slot="logo" class="">
+        <div class="flex title h48 pl20">
+          <h1 class="type-title center">Rapidocus</h1>
+        </div>
+      </div>
+      <div slot="nav-logo"></div>
+    </rapi-doc>
 
     <template #sidebar>
       <sidebar-detail icon="settings" title="General settings" close>
         <div class="layout-options">
-          <div class="type-label">Display mode</div>
-          <v-select v-model="display_mode" :items="display_modes"> </v-select>
+          <div class="pb20">
+            <div class="type-label">Display mode</div>
+            <v-select v-model="display_mode" :items="display_modes"> </v-select>
+          </div>
+        </div>
+        <v-checkbox v-model="allow_try">allow_try</v-checkbox>
+      </sidebar-detail>
+      <sidebar-detail icon="top_panel_close" title="Header">
+        <div class="layout-options">
+          <v-checkbox v-model="show_header">Show header</v-checkbox>
+          <v-checkbox v-model="allow_spec_url_load"
+            >Allow spec url load</v-checkbox
+          >
+          <v-checkbox v-model="allow_spec_file_load"
+            >Allow spec file load</v-checkbox
+          >
         </div>
       </sidebar-detail>
-      <sidebar-detail icon="layers" title="Navigation bar">
+      <sidebar-detail icon="left_panel_close" title="Navigation">
         <div class="layout-options">
           <div class="pb20">
             <div class="type-label">Show method in navbar</div>
@@ -112,12 +152,34 @@ const colors = computed(() => {
             >
             </v-select>
           </div>
-          <div class="type-label">Navbar active item</div>
-          <v-select
-            v-model="nav_active_item_marker"
-            :items="nav_active_items_marker"
+          <div class="pb20">
+            <div class="type-label">Navbar active item</div>
+            <v-select
+              v-model="nav_active_item_marker"
+              :items="nav_active_items_marker"
+            >
+            </v-select>
+          </div>
+          <div class="pb20">
+            <div class="type-label">Font spacing</div>
+            <v-select v-model="font_spacing" :items="font_spacings"> </v-select>
+          </div>
+          <div class="pb20">
+            <div class="type-label">Font size</div>
+            <v-select v-model="font_size" :items="font_sizes"> </v-select>
+          </div>
+          <v-checkbox v-model="show_info">Show info</v-checkbox>
+          <v-checkbox v-model="allow_server_selection"
+            >Allow server selection</v-checkbox
           >
-          </v-select>
+          <v-checkbox v-model="allow_authentication"
+            >Allow authentication</v-checkbox
+          >
+          <v-checkbox v-model="persist_auth">Persist auth</v-checkbox>
+          <v-checkbox v-model="allow_search">Allow search</v-checkbox>
+          <v-checkbox v-model="allow_advanced_search"
+            >Allow advanced search</v-checkbox
+          >
         </div>
       </sidebar-detail>
     </template>
@@ -125,10 +187,25 @@ const colors = computed(() => {
 </template>
 
 <style>
+.flex {
+  display: flex;
+}
+.center {
+  align-content: center;
+}
+.h48 {
+  height: 48px;
+}
 .pb20 {
   padding-bottom: 20px;
 }
+.pl20 {
+  padding-left: 20px;
+}
 .api-rapidocus #navigation .module-nav {
+  display: none !important;
+}
+.api-rapidocus .header-bar {
   display: none !important;
 }
 </style>